context-webrtc-be/final_requirements.txt
pydub = "*"
aiortc = "*"
ffmpeg-python = "*"
ffmpeg = "*"
python-socketio = "*"
av = "*"
speechrecognition = "*"
transformers = "*"
google-genai = "*"
python-dotenv = "*"